feat(#81): delete constraint propagate move to delete constraint and refactor
